Added preliminary support for new "states" attribute to <query states="xxx"> element, where the value can be: - "none":  Do not apply the query to states. (This is the default for backward compatibility.) - "together" : Apply the query to states and global regions. - "only": Apply the query to states only, excluding global regions (Region "USA" is also excluded.)
Use of this feature currently requires that the .pygcam.cfg file define a new variable, "Test.RegionDiscovery = True".  This requirement will be removed once testing is complete.

Note: Please copy the file pygcam/pygcam/etc/project-schema.xml to the "etc" directory in your project folder(s). It defines the legal syntax for the project.xml file and includes the new attribute. This really only matters if you're using an XML editing app that can read the schema (which I recommend!) since this allows the app to detect syntax errors and to offer hints.